backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.12.5
pydantic-settings==2.7.0
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
pandas==3.0.2
pydantic-ai-slim[google]==1.82.0
